requirements.txt
PySide6>=6.7
PyPDF2>=3.0
python-docx>=0.8.11
pyspellchecker>=0.8.1
